feat(scan): allow decimal start frequency (e.g. 400.2)
startfreq was an int config (type 0 / atoi), so a value like 400.2 was silently
truncated to 400 and the scan/spectrum could only start on whole-MHz boundaries.
Make it a double using the existing -7 (double) config type, like rxlat/rxlon.

The scanner sweep, the on-device spectrum axis labels (now formatted with %g),
and the scan-plot JSON (%.5g) all handle decimals; a blank/NaN field falls back
to 400 MHz. Lets the user move the ~6 MHz sweep window up to skip known low-end
birdies (e.g. start at 400.2 to avoid ~400.0-400.11 spurs).

+++ RX_FSK/src/Sonde.h
@@ -303,7 +303,7 @@ typedef struct st_rdzconfig {
 	int dispsaver;			// Turn display on/off (0=always on, 10*n+1: off after n seconds, 
 					//	10*n+2: scanner off after n seconds, RX always shown)
 	int dispcontrast;		// For OLED: set contrast to 0..255 (-1: don't set/leave at factory default)
-	int startfreq;			// spectrum display start freq (400, 401, ...)
+	double startfreq;		// spectrum/scan start freq in MHz (decimals allowed, e.g. 400.2); sweep covers startfreq..startfreq+~6 MHz
 	int channelbw;			// spectrum channel bandwidth (valid: 5, 10, 20, 25, 50, 100 kHz)	
 	int spectrum;			// show freq spectrum for n seconds -1=disable; 0=forever
 	int marker;				// show freq marker in spectrum  0=disable
